Refactor notification settings into feature module
