+++ android/lib/feature/notification/impl/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    alias(libs.plugins.mullvad.android.library)
+    alias(libs.plugins.mullvad.android.library.feature.impl)
+    alias(libs.plugins.mullvad.android.library.compose)
+    alias(libs.plugins.kotlin.android)
+    alias(libs.plugins.kotlin.parcelize)
+    alias(libs.plugins.kotlin.ksp)
+}
+
+android {
+    namespace = "net.mullvad.mullvadvpn.feature.notification.impl"
+    ksp { arg("compose-destinations.moduleName", "notification") }
+}
+
+dependencies {
+    implementation(projects.lib.repository)
+
+    implementation(libs.koin.compose)
+    implementation(libs.arrow)
+    implementation(libs.protobuf.kotlin.lite)
+
+    // Destinations
+    implementation(libs.compose.destinations)
+    ksp(libs.compose.destinations.ksp)
+}
