+++ android/settings.gradle.kts
@@ -49,6 +49,7 @@ include(
     ":lib:feature:filter:impl",
     ":lib:feature:managedevices:impl",
     ":lib:feature:multihop:impl",
+    ":lib:feature:notification:impl",
     ":lib:feature:problemreport:impl",
     ":lib:feature:redeemvoucher:impl",
     ":lib:feature:serveripoverride:impl",
